grocery_list_calculator.py
This is a grocery list calculator, which uses a split input function combined with looping to collect two sets of data: The items being purchased and the cost of these items. the program then tallys up the total before tax, shows the local and state sales tax totals, and then the total with state and local taxes added.